requirements.txt
numpy==1.26.4
svgwrite==1.4.3
Flask==3.0.3
Flask-Cors==4.0.1
Pillow==10.0.0
cairosvg==2.7.1
# 强制安装CairoSVG依赖，避免栅格化失败回退为矢量
cairocffi==1.7.0
tinycss2==1.2.1
cssselect2==0.7.0
defusedxml==0.7.1
# 标点符号系统依赖
fonttools==4.47.0
